{{cookiecutter.project_directory}}/requirements.txt
alembic==1.8.0
blinker==1.4
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
gunicorn==20.0.4
